+++ Drakengard1and2Extractor/FileExtraction/FileKPS.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Drakengard1and2Extractor.FileExtraction
+{
+    internal class FileKPS
+    {
+
+    }
+}
