+++ Drakengard1and2Extractor/ImageConversion/FileSPK0.resx
