+++ Drakengard1and2Extractor/Support/Extensions/StreamHelpers.cs
