+++ Drakengard1and2Extractor/Support/ImageHelpers/BmpPngHelpers.cs
