+++ Drakengard1and2Extractor/Support/ImageHelpers/ImgOptions.cs
